We have some working conversions!
Request for conversion to Gregorian of a date "1594-11-20" for location
search terms "groningen".  Finds the period where Julian calendar was
re-activated and also hints there are multiple calendars matching
"Groningen".

$ curl -s localhost:8080/convert -H 'Content-type: application/json' -d '{"year": "1594", "month": "11", "day": "20", "location": "groningen", "type": "gregorian"}' | jq .
{
  "date": {
    "year": 1594,
    "month": 11,
    "day": 30,
    "notes": [
      "Date on or before end of 'Julian' calendar"
    ]
  },
  "hints": [
    "Multiple calendars found for 'groningen', retry with a specific one for greater accuracy: [\"Dutch Republic, Friesland and Groningen\", \"Groningen (City)\"]"
  ]
}

The result currently only contains the first succesful conversion, but
multiple are computed, todo: return all and / or score them in a way.